Fix referencing_columns for array values
It's not as simple as just looking for arrays, since we are dealing with
possibly multi-column keys, they are all arrays.

This change means if we have "5 7 9" as the column, then it's as if we
had three separate rows with "5", "7", and "9".

The checking code now does not need to change.